Projet_master_RNCP/project_master_ETL/requirements.txt
# python_version recommanded: 3_11
flask
flask_cors
# werkzeug==2.3.7
Flask-APScheduler
python-dotenv
pandas
pymongo
requests
scipy
openpyxl